pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "qlibworks"
version = "0.1.0"
description = "Qlib research workspace wired to DuckDB data source"
requires-python = ">=3.8"
dependencies = [
  "pandas>=1.5",
  "numpy>=1.23",
  "duckdb>=0.9.0",
  "joblib>=1.3",
  "PyYAML>=6.0",
  "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
]
research = [
  "pyqlib>=0.9.0",
  "scikit-learn>=1.2",
  "scipy>=1.10",
  "matplotlib>=3.7",
]
dl = [
  "torch>=2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "S", "NPY", "PD"]
ignore = ["E501"]
